% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projScaffold.R
\name{projectScaffold}
\alias{projectScaffold}
\title{A Project Scaffolding Function}
\usage{
projectScaffold()
}
\description{
This function builds a basic project scaffold within the working directory
}
\examples{
projScaffold()
}
\keyword{project}
\keyword{scaffold}
